requirements.txt
numpy
pyyaml
shapely
pygame